+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add support for Rails 7.0 and Ruby 3.1 (no changes required)
 * Fix escaping issue on boolean input with `include_hidden: false` and custom wrapper.
 * Update Bootstrap install generator version 5. [@mhw](https://github.com/mhw)
+* Accept proc as `group_method` for grouped collection select
 
 ## 5.1.0