+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Update Bootstrap install generator version 5. [@mhw](https://github.com/mhw)
 * Accept proc as `group_method` for grouped collection select
 * Honor `include_hidden` option on inline boolean inputs [@yboulkaid](https://github.com/yboulkaid)
+* Fix deprecation error when using country_select input.
 
 ## 5.1.0